Updating LoginActivity and MyHomeActivity
