+++ app/src/main/res/layout/activity_my_home.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="cmput301f17t01.bronzify.activities.MyHomeActivity">
+    tools:context="cmput301f17t01.bronzify.activities.LoginActivity">
 
     <TextView
         android:layout_width="wrap_content"
